requirements.txt
protobuf>=4.25.8
requests==2.32.4
fastapi~=0.115.8
uvicorn~=0.34.0
pydantic~=2.10.6
python-dotenv~=1.0.1
pytz~=2025.1
asyncpg~=0.30.0
APScheduler~=3.11.0
loguru~=0.7.3
beautifulsoup4~=4.13.3
tenacity~=9.0.0
aiohttp~=3.11.0